Fixes #167 - Added Favicon and static app bar and fixed topbar and button colors
