+++ src/components/Auth/Login/Login.css
@@ -1,17 +1,29 @@
 .forgotpwdlink{
-    text-decoration: none;
-    color: #999999;
-    margin-bottom: 20px;
+	color: #999999;
+	cursor: pointer;
 }
+
+.forgotpwdlink:hover{
+    color: #039be5;
+}
+
 .loginForm{
-    margin: 0 auto;
-    background: #ffffff;
+	margin: 0 auto;
+	padding: 10px 0;
 }
+
 .sty{
-    width: 256px;
+	width: 256px;
 }
+
+.loginForm h3{
+	font-family: 'Open Sans', sans-serif;
+	margin: 5px 0;
+	font-weight: 500;
+}
+
 @media screen and (max-width: 768px) {
-    .loginForm h1{
-        font-size: 16px;
-    }
+	.loginForm h3{
+		font-size: 16px;
+	}
 }
