+++ src/components/Auth/Login/Login.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import Paper from 'material-ui/Paper';
 import TextField from 'material-ui/TextField';
 import RaisedButton from 'material-ui/RaisedButton';
-import { Link } from 'react-router-dom';
 import './Login.css';
 import PasswordField from 'material-ui-password-field'
 import $ from 'jquery';
@@ -184,7 +183,6 @@ class Login extends Component {
 
     render() {
         // const { token } = this.props;
-        const hidden = '';
 
         const styles = {
             'width': '100%',
@@ -202,9 +200,8 @@ class Login extends Component {
                 </div>
                 <div className="loginForm">
                     <Paper zDepth={0}style={styles}>
-                        <h1>Login to SUSI</h1>
+                        <h3>Login to SUSI</h3>
                         <form onSubmit={this.handleSubmit}>
-
                             <div>
                                 <TextField name="email"
                                            value={this.state.email}
@@ -217,39 +214,38 @@ class Login extends Component {
                                     name='password'
                                     style={fieldStyle}
                                     value={this.state.password}
-
                                     onChange={this.handleChange}
                                     errorText={this.passwordErrorMessage}
                                     floatingLabelText='Password' />
                             </div>
-                            <div>
-                                {hidden}
-                            </div>
+                            <span style={{
+                              margin: '3px 0'
+                            }}>{this.state.msg}</span>
                             <div>
                                 <RaisedButton
                                     label="Login"
                                     type="submit"
-                                    backgroundColor="#607D8B"
+                                    backgroundColor="#4285f4"
                                     labelColor="#fff"
-                                    disabled={!this.state.validForm} />
+                                    disabled={!this.state.validForm}
+                                    style={{margin:'15px 0 '}}/>
                             </div>
-                            <span>{this.state.msg}</span>
-                            <h1>OR</h1>
-                            <div>
-                                <Link to='/forgotpwd'
-                                      className="forgotpwdlink">
-                                    <b>Forgot Password?</b>
-                                </Link>
-                            {/*</div>*/}
-
-                             </div>
+                            <span className="forgotpwdlink"
+                              onClick={this.handleForgotPassword}>
+                              Forgot Password?
+                            </span>
+                            <br />
+                            <h4 style={{
+                              margin: '7px 0'
+                            }}>OR</h4>
                             <div>
                                 <h4>If you do not have an account, Please SignUp</h4>
                                     <RaisedButton
                                         label='SignUp'
                                         onTouchTap={this.handleSignUp}
-                                        backgroundColor="#19314B"
-                                        labelColor="#fff" />
+                                        backgroundColor="#4285f4"
+                                        labelColor="#fff"
+                                        style={{margin:'15px 0 0 0'}}/>
                             </div>
                         </form>
                     </Paper>
