+++ src/components/NotFound/NotFound.js
@@ -1,16 +1,14 @@
-
 import React from 'react';
-
+import StaticAppBar from '../StaticAppBar/StaticAppBar.react';
 
 export default class NotFound extends React.Component {
-
 
     render() {
         return(
             <div>
+                <StaticAppBar {...this.props} />
                 <h1 style={styles.bg}>404 &nbsp; :(</h1>
             </div>
-
         )
     }
 }
@@ -21,6 +19,6 @@ const styles = {
         lineHeight:'90px',
         textAlign: 'center',
         fontSize:'200px',
-        marginLeft:'80px'
+        padding: "100px 30px 30px",
     }
 }
