+++ src/components/Settings/Settings.js
@@ -1,13 +1,13 @@
 import React from 'react';
-
+import StaticAppBar from '../StaticAppBar/StaticAppBar.react';
 
 export default class Settings extends React.Component {
-
 
     render() {
         return(
             <div>
-                <h1 style={styles.bg}>Settings</h1>
+              <StaticAppBar {...this.props} />
+              <h1 style={styles.bg}>Settings</h1>
             </div>
 
         )
@@ -19,6 +19,6 @@ const styles = {
         height: '90px',
         lineHeight:'90px',
         textAlign: 'center',
-        marginLeft:'80px'
+        padding: "80px 30px 30px",
     }
 }
