+++ src/components/SkillEditor/SkillEditor.js
@@ -19,6 +19,7 @@ import 'brace/theme/solarized_light';
 import 'brace/theme/terminal';
 import * as $ from "jquery";
 import colors from "../../Utils/colors";
+import StaticAppBar from '../StaticAppBar/StaticAppBar.react';
 
 const models = [];
 const groups = [];
@@ -150,6 +151,8 @@ export default class Container extends React.Component {
             padding: "10px"
         };
         return (
+          <div>
+            <StaticAppBar {...this.props} />
             <div style={styles.home}>
                 <Paper style={style} zDepth={1}>
                     <div style={styles.center}>
@@ -224,14 +227,15 @@ export default class Container extends React.Component {
                     />
                 </div>
             </div>
-
+          </div>
         );
     }
 }
 
 const styles = {
     home: {
-        width: '100%'
+        width: '100%',
+        padding: "80px 30px 30px",
     },
     center: {
         display: "flex",
