+++ src/components/StaticAppBar/StaticAppBar.css
@@ -1,0 +1,44 @@
+header{
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 1000;
+  height: 46px;
+  width: 100vw;
+  background: #607d8b;
+  box-shadow: 0 2px 2px 0 rgba(0,0,0,.14), 0 3px 1px -2px rgba(0,0,0,.2), 0 1px 5px 0 rgba(0,0,0,.12);
+  user-select: none;
+  -webkit-user-select: none;
+}
+
+header {
+    height: 46px;
+    position: fixed;
+    top: 0;
+    transition: top 0.9s ease-in-out;
+    width: 100%;
+}
+
+.nav-down{
+    box-shadow: 0 0 4px rgba(0,0,0,.14), 0 4px 8px rgba(0,0,0,.28) !important;
+}
+
+.nav-up {
+    top: -100px;
+}
+
+.siteTitle {
+    height: 25px;
+    margin-top: 21px;
+    display: block;
+}
+
+.topAppBar button{
+    display: none !important;
+}
+
+.topAppBar div div div button {
+    display: block !important;
+    visibility: visible;
+    box-shadow: none;
+}
