+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 // import registerServiceWorker from './registerServiceWorker';
-import Header from './components/Header/Header';
 import Home from "./components/SkillEditor/SkillEditor";
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import { BrowserRouter as Router, Route} from "react-router-dom";
@@ -11,7 +10,6 @@ import Switch from "react-router-dom/es/Switch";
 import BrowseHistory from "./components/BrowseHistory/BrowseHistory";
 import BrowseSkill from "./components/BrowseSkill/BrowseSkill";
 import injectTapEventPlugin from 'react-tap-event-plugin';
-import Body from "./components/Body/Body";
 import BrowseExamples from "./components/BrowseExamples/BrowseExamples";
 import VisualEditor from "./components/VisualEditor/VisualEditor";
 import SkillListing from "./components/SkillPage/SkillListing";
@@ -20,43 +18,25 @@ injectTapEventPlugin();
 class App extends React.Component {
 
     render() {
+        document.body.style.backgroundColor = "#eee";
         return (
-            <Router>
-        <MuiThemeProvider>
-
-            <div style={styles.app}>
-                <Header />
-                <Body>
-                    <Switch>
-                        <Route path="/skillEditor" component={Home} />
-                        <Route path="/browseHistory" component={BrowseHistory} />
-                        <Route path="/browseExamples" component={BrowseExamples} />
-                        <Route path="/browseSkill" component={BrowseSkill}/>
-                        <Route path="/visualEditor" component={VisualEditor}/>
-                        <Route path="/skillPage" component={SkillListing}/>
-                        <Route path="/settings" component={Settings}/>
-                        <Route path="/" component={BrowseSkill} />
-                        <Route component={NotFound} />
-                    </Switch>
-                </Body>
-            </div>
-        </MuiThemeProvider>
-            </Router>
+          <Router>
+            <MuiThemeProvider>
+              <Switch>
+                <Route exact path="/skillEditor" component={Home} />
+                <Route exact path="/browseHistory" component={BrowseHistory} />
+                <Route exact path="/browseExamples" component={BrowseExamples} />
+                <Route exact path="/browseSkill" component={BrowseSkill}/>
+                <Route exact path="/visualEditor" component={VisualEditor}/>
+                <Route exact path="/skillPage" component={SkillListing}/>
+                <Route exact path="/settings" component={Settings}/>
+                <Route exact path="/" component={BrowseSkill} />
+                <Route exact path="*" component={NotFound} />
+              </Switch>
+            </MuiThemeProvider>
+          </Router>
         );
     }
 }
-
-const styles = {
-    app: {
-        width: '100%',
-        height: '100%',
-        display: 'flex',
-        alignItems: 'stretch',
-        justifyContent: 'stretch',
-        background: '#eee',
-        flexDirection: "row",
-        overflow: "hidden"
-    }
-};
 
 ReactDOM.render(<App />, document.getElementById('root'));
